✨ [Texture]: Allow texture specific UV wrapping modes
UV wrapping modes used to default to REPEAT. However, this would cause
artefacts when sampling textures magnified using the nearest neighbour
algorithm. Changing the default to CLAMP would break sampling outside
of the [0,1] range. To resolve this a new function has been added that
allows the user to specify on a per texture basis the UV wrapping to
use on both the X and Y axis.