✨ [Logger]: Custom logger for PyASGE
ASGE uses a logger that has a number of predefined styles and
levels. The "loggable" level is set by the build type i.e. Release
or Debug. As the DLL bundled with PyASGE is a Release build, any
calls to the logger above ERROR are automatically discared. To
circumvent this a custom logger has been added to PyASGE, that does
not obey the level limits. This allows for easy logging to be done
in game and to the std out with the minimum of fuss.