✨ [Game]: Introduction of Unity style fixed updates
The update function has been amended to be variable instead of
fixed. A new function named `fixed_update` has been added which
acts in a typical fixed timestep approach. The game will always
pass through the same amount of game time delta to this function.
This is perfect for determinstic operations i.e. physics or
network synchronisation.

The original update function can now be thought of as a frame
update. The game will attempt to render as quickly as possible (up to
the specified FPS limit) and the update function will always run
only once before the screen is redrawn. This is contrast to
`fixed_update` which will attempt to catch-up when falling behind.

If things get too gnarly, the game will cap the amount of game time
being simulated and everything will slow down. For this to happen
the frame time has to be very poor.