✨ [RenderTarget]: Support for MSAA render targets
As the render targets for ASGE have been changed to make use of
multi-sampling, some small changes were required to the bindings.
The most relevant is that of the function call `resolve`. This is
used to convert the multi-sampled textures in to a standard 2D
texture, suitable for attachment in a shader. At present there is
no way to manually specify the sampling levels, though it can be
turned off globally using `GameSettings`.