✨ [GamePad]: Support for gamepad lists
It is now possible to retrieve all connected gamepads as a list.
This makes it easier to iterate over them and can be used for
storing state updates or even just for checking input values. It's
still possible to retrieve individual gamepads using the `getGamePad`
function. If no index is specified, the first connected is returned.
This could be used to quickly check if a gamepad is still connected
or if it has been suddenly disconnected.